+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.1.0
 * added `BackupShell::send()` and `BackupManager::send()` methods to send backup
     files via mail;
+* added `--send` option to `BackupShell::export()`;
 * added `BackupExport::send()` to set up the email recipient, so by calling the
     `export()` method the backup file will be sent via email;
 * added `BackupTrait`,  that provides some methods used by all other classes;